--- some changes.(again)

+++ src/main/java/com/example/controller/UserController.java
@@ -56,17 +56,16 @@ public class UserController {
 
     @DeleteMapping("/users/{id}")
     public void deleteUser(@PathVariable int id) {
-        User user = service.deleteById(id);
 
-        if (user == null)
-            throw new UserNotFoundException("id-" + id);
+        if (service.exists(id)) {
+            service.deleteById(id);
+        } else {
+            throw new UserNotFoundException("id = " + id);
+        }
+
+
     }
 
-    //
-    // input - details of user
-    // output - CREATED & Return the created URI
-
-    //HATEOAS
 
     @PostMapping("/users")
     public ResponseEntity<Object> createUser(@Valid @RequestBody User user) {
